+++ pkg/frontend/querymiddleware/stats.go
@@ -159,9 +159,6 @@ type QueryDetails struct {
 
 	ResultsCacheMissBytes int
 	ResultsCacheHitBytes  int
-	// SamplesProcessedCacheAdjusted represents the total number of samples processed by queriers to produce the query result.
-	// It includes samples has been stored in the query-frontend cache and then fetched to produce the current full query result.
-	SamplesProcessedCacheAdjusted uint64
 }
 
 type contextKey int